requirements.txt
transformers
datasets
torch
langchain
langchain-community
faiss-cpu
pypdf
sentence-transformers
nltk
langchain-huggingface